pyproject.toml
[project]
name = "facialrecognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.35",
]
